# api/src/products-from-orders/dto/update-products-from-order.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateProductsFromOrderDto } from './create-products-from-order.dto';

export class UpdateProductsFromOrderDto extends PartialType(CreateProductsFromOrderDto) {}