# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/product":{"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["product"]},"get":{"operationId":"ProductController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["product"]}},"/product/generate":{"post":{"operationId":"ProductController_generateStock","parameters":[],"responses":{"201":{"description":""}},"tags":["product"]}},"/product/{id}/increase/{quantity}":{"patch":{"operationId":"ProductController_increaseQuantity","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"quantity","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["product"]}},"/product/{id}/decrement/{quantity}":{"patch":{"operationId":"ProductController_decrementQuantity","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"quantity","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["product"]}},"/product/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["product"]},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["product"]},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["product"]}},"/order":{"post":{"operationId":"OrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["order"]},"get":{"operationId":"OrderController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["order"]}},"/order/{id}":{"get":{"operationId":"OrderController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["order"]},"patch":{"operationId":"OrderController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":""}},"tags":["order"]},"delete":{"operationId":"OrderController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["order"]}},"/stock":{"post":{"operationId":"StockController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStockDto"}}}},"responses":{"201":{"description":""}},"tags":["stock"]},"get":{"operationId":"StockController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["stock"]},"delete":{"operationId":"StockController_removeAll","parameters":[],"responses":{"200":{"description":""}},"tags":["stock"]}},"/stock/generate":{"post":{"operationId":"StockController_generateStock","parameters":[],"responses":{"201":{"description":""}},"tags":["stock"]}},"/stock/{id}":{"get":{"operationId":"StockController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["stock"]},"patch":{"operationId":"StockController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStockDto"}}}},"responses":{"200":{"description":""}},"tags":["stock"]},"delete":{"operationId":"StockController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["stock"]}},"/customer":{"post":{"operationId":"CustomerController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomerDto"}}}},"responses":{"201":{"description":""}},"tags":["customer"]},"get":{"operationId":"CustomerController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["customer"]}},"/customer/{id}":{"get":{"operationId":"CustomerController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["customer"]},"patch":{"operationId":"CustomerController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerDto"}}}},"responses":{"200":{"description":""}},"tags":["customer"]},"delete":{"operationId":"CustomerController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["customer"]}},"/product-from-orders":{"post":{"operationId":"ProductsFromOrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductsFromOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["product from orders"]},"get":{"operationId":"ProductsFromOrderController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["product from orders"]}},"/product-from-orders/{id}":{"get":{"operationId":"ProductsFromOrderController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["product from orders"]},"patch":{"operationId":"ProductsFromOrderController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductsFromOrderDto"}}}},"responses":{"200":{"description":""}},"tags":["product from orders"]},"delete":{"operationId":"ProductsFromOrderController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["product from orders"]}}},"info":{"title":"palamed-api","description":"The palamed-api description","version":"1.3.5","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateProductDto":{"type":"object","properties":{"reference":{"type":"number"},"label":{"type":"string"},"quantity":{"type":"number"},"type":{"type":"string"}},"required":["reference","label","quantity","type"]},"UpdateProductDto":{"type":"object","properties":{}},"Customer":{"type":"object","properties":{"id":{"type":"number"}},"required":["id"]},"CreateOrderDto":{"type":"object","properties":{"reference":{"type":"number"},"customer":{"$ref":"#/components/schemas/Customer"},"productsFromOrder":{"type":"array","items":{"type":"string"}}},"required":["reference","customer","productsFromOrder"]},"UpdateOrderDto":{"type":"object","properties":{}},"CreateStockDto":{"type":"object","properties":{"id":{"type":"number"},"capacity":{"type":"number"},"product":{"type":"array","items":{"type":"string"}}},"required":["id","capacity","product"]},"UpdateStockDto":{"type":"object","properties":{}},"CreateCustomerDto":{"type":"object","properties":{"id":{"type":"number"},"lastname":{"type":"string"},"firstname":{"type":"string"},"order":{"type":"array","items":{"type":"string"}}},"required":["id","lastname","firstname","order"]},"UpdateCustomerDto":{"type":"object","properties":{"id":{"type":"number"},"lastname":{"type":"string"},"firstname":{"type":"string"},"order":{"type":"array","items":{"type":"string"}}}},"Order":{"type":"object","properties":{"reference":{"type":"number"},"customer":{"$ref":"#/components/schemas/Customer"}},"required":["reference","customer"]},"Product":{"type":"object","properties":{"reference":{"type":"number"}},"required":["reference"]},"CreateProductsFromOrderDto":{"type":"object","properties":{"productsFromOrderId":{"type":"number"},"order":{"$ref":"#/components/schemas/Order"},"product":{"$ref":"#/components/schemas/Product"},"quantity":{"type":"number"}},"required":["productsFromOrderId","order","product","quantity"]},"UpdateProductsFromOrderDto":{"type":"object","properties":{"productsFromOrderId":{"type":"number"},"order":{"$ref":"#/components/schemas/Order"},"product":{"$ref":"#/components/schemas/Product"},"quantity":{"type":"number"}}}}}}